+++ tests/common/HtmlTest.php
@@ -847,6 +847,15 @@ final class HtmlTest extends TestCase
                 ],
             ],
             [
+                '',
+                [
+                    'data-test1' => null,
+                    'data' => [
+                        'test2' => null,
+                    ],
+                ],
+            ],
+            [
                 ' src="xyz" data-a="1" data-b="c"',
                 ['src' => 'xyz', 'data' => ['a' => 1, 'b' => 'c']],
             ],