+++ src/Html.php
@@ -7,7 +7,6 @@ namespace Yiisoft\Html;
 use InvalidArgumentException;
 use JsonException;
 use Stringable;
-use ValueError;
 use Yiisoft\Html\Tag\A;
 use Yiisoft\Html\Tag\Address;
 use Yiisoft\Html\Tag\Article;
@@ -1957,12 +1956,12 @@ final class Html
      * Render attribute in HTML tag.
      *
      * @link https://html.spec.whatwg.org/#a-quick-introduction-to-html
-     * @throws ValueError When name contains any of the following: SPACE " ' > / =
+     * @throws InvalidArgumentException When name contains any of the following: SPACE " ' > / =
      */
     private static function renderAttribute(string $name, string $encodedValue = '', string $quote = '"'): string
     {
         if (preg_match('/[\s"\'>\/\=]/', $name) > 0) {
-            throw new ValueError(
+            throw new InvalidArgumentException(
                 'Name "' . $name . '" contains invalid character(s). Attribute name must not contain: [SPACE, ", \', >, /, =]'
             );
         }