+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## 3.5.1 under development
 
-- Bug #209: Fix output of `null` value attributes in `Html::renderTagAttributes()` when attribute is in `Html::DATA_ATTRIBUTES` or `Html::ATTRIBUTES_WITH_CONCATENATED_VALUES` (@es-sayers)
-
+- Bug #208: Fix output of `null` value attributes in `Html::renderTagAttributes()` (@es-sayers)
 
 ## 3.5.0 July 11, 2024