+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 3.5.1 under development
 
-- Bug #212: Add `ValueError Exception` in `Html::renderAttribute()` for illegal characters in `$name` (@es-sayers)
+- Bug #212: Add `InvalidArgumentException` in `Html::renderAttribute()` for illegal characters in `$name` (@es-sayers)
 - Bug #208: Fix output of `null` value attributes in `Html::renderTagAttributes()` (@es-sayers)
 - Enh #214: Add `Stringable` and array values support to textarea tag (@vjik)